requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
certifi==2024.2.2
frozenlist==1.4.1
idna==3.7
lxml==5.2.2
magic-filter==1.0.12
multidict==6.0.5
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.1
pydantic_core==2.18.2
PyMuPDF==1.24.3
PyMuPDFb==1.24.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.1
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
yarl==1.9.4
